Reviewed the types
